Added tooltip ability for structure properties Added float to packets Changed forest rate for different forest levels Changed lumbermill abilities Fixed troop crashing bug when stub was null and a get troop info came in Hopefully fixed scheduler bug Relaxed attack points. Will gain 1 attack point per unit killed. Changed battle turn to 15 seconds per hit Forest sends system message to player when it's depleted completely

+++ Game/Data/ForestManager.cs
@@ -28,11 +28,11 @@ namespace Game.Data {
             forests.Add(forest.ObjectId, forest);
         }
 
-        public void CreateForest(byte lvl, int capacity, int rate) {
+        public void CreateForest(byte lvl, int capacity, double rate) {
             CreateForestAt(lvl, capacity, rate, 0, 0);
         }
 
-        public void CreateForestAt(byte lvl, int capacity, int rate, uint x, uint y) {
+        public void CreateForestAt(byte lvl, int capacity, double rate, uint x, uint y) {
             Forest forest = new Forest(lvl, capacity, rate);
 
             if (x == 0 || y == 0) {
